% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierpr.R
\encoding{UTF-8}
\name{hierpr}
\alias{hierpr}
\title{Hierarchical precision}
\usage{
hierpr(truth, response, type = "micro")
}
\arguments{
\item{truth}{True (observed) labels. Must have the same length as \code{response}. Can be factor (recommended) or character.
These need to be structured in the same way as required for \code{\link{topdown}}. See the 'Details' section of \code{\link{topdown}}
for details.}

\item{response}{Predicted response labels. Must have the same length as \code{truth}. Can be factor (recommended) or character.
These need to be structured in the same way as required for \code{\link{topdown}}. See the 'Details' section of \code{\link{topdown}}
for details.}

\item{type}{Type of average to use: The micro-average (\code{type="micro"}) and the macro-average (\code{type="macro"}) give
the same weight to all observations and all classes, respectively. Default is "micro".}
}
\value{
The value of the hierarchical precision.
}
\description{
Hierarchical precision defined by Kiritchenko et al. (2005). 
Takes the ancestors of the true and predicted classes into account, but has the disadvantage that it overpenalizes errors in nodes with many
ancestors (Kosmopoulos et al., 2015).
}
\examples{
\dontrun{

## Load package:

library("hierclass")


## Load the example data set 'datasim':

data(datasim)


## Set seed to make results reproducible:

set.seed(1234)


## Split data set into training and test data:

trainind <- sample(1:nrow(datasim), size=round((3/4)*nrow(datasim)))
datatrain <- datasim[trainind,]
datatest <- datasim[-trainind,]


## Construct a top-down hierarchical prediction rule using the training data:

object <- topdown(ydepvar ~ ., data=datatrain, num.trees=50)
# NOTE: In practice 'num.trees' should in general be larger
# to achieve better performance (default is 500).
# We use 50 trees here only for computational efficiency of
# the example.


## Predict the classes of observations in the test data (without
## early stopping because 'confid=1' by default):

preds <- predict(object, data=datatest)


## Compute the micro- and macro-average hierarchical precision
## for the predictions of the test data:

hierpr(truth=datatest$ydepvar, response=preds)
hierpr(truth=datatest$ydepvar, response=preds, type="macro")
}

}
\references{
\itemize{
  \item Kiritchenko, S., Matwin, S., Famili, A. F. (2005). Functional annotation of genes using hierarchical text categorization. In: Proceedings of the BioLINK SIG: Linking Literature, Information and Knowledge for Biology, Detroit, Michigan, USA.
  \item Kosmopoulos, A., Partalas, I., Gaussier, E., Paliouras, G., Androutsopoulos, I. (2015). Evaluation measures for hierarchical classification: A unified view and novel approaches. Data Mining and Knowledge Discovery 29(3):820–865, <\doi{10.1007/s10618-014-0382-x}>.
  }
}
\seealso{
\code{\link{hierre}}, \code{\link{hierfbeta}}, \code{\link{hloss}}, \code{\link{spath}}
}
\author{
Roman Hornung
}
