% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr3_measures_hierclass.R
\name{mlr_measures_classif.hierpr}
\alias{mlr_measures_classif.hierpr}
\title{Hierarchical precision}
\description{
Hierarchical precision defined by Kiritchenko et al. (2005). 
Takes the ancestors of the true and predicted classes into account, but has the disadvantage that it overpenalizes errors in nodes with many
ancestors (Kosmopoulos et al., 2015).
}
\note{
The score function calls \code{\link[=hierpr]{hierpr()}} from package \CRANpkg{hierclass}.

If the measure is undefined for the input, \code{NaN} is returned.
This can be customized by setting the field \code{na_value}.
}
\section{Dictionary}{

This \link{Measure} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary} \link{mlr_measures} or with the associated sugar function \code{\link[=msr]{msr()}}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{mlr_measures$get("classif.hierpr")
msr("classif.hierpr")
}\if{html}{\out{</div>}}
}

\section{Parameters}{
\tabular{llll}{
   Id \tab Type \tab Default \tab Levels \cr
   type \tab character \tab - \tab micro, macro \cr
}
}

\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[=Measure]{Measures}: \link{mlr_measures}

\code{as.data.table(mlr_measures)} for a complete table of all (also dynamically created) \link{Measure} implementations.

Other classification measures: 
\code{\link{mlr_measures_classif.hierfbeta}},
\code{\link{mlr_measures_classif.hierre}},
\code{\link{mlr_measures_classif.hloss}},
\code{\link{mlr_measures_classif.spath}}

Other binary classification measures: 
\code{\link{mlr_measures_classif.hierfbeta}},
\code{\link{mlr_measures_classif.hierre}},
\code{\link{mlr_measures_classif.hloss}},
\code{\link{mlr_measures_classif.spath}}
}
\concept{binary classification measures}
\concept{classification measures}
