% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr3_measures_hierclass.R
\docType{data}
\name{mlr_measures_classif.hloss}
\alias{mlr_measures_classif.hloss}
\alias{measure_hloss}
\title{H-loss measure}
\format{
An object of class \code{MeasureClassifHierarchical} (inherits from \code{MeasureClassif}, \code{Measure}, \code{R6}) of length 25.
}
\usage{
measure_hloss
}
\description{
H-loss measure defined by Cesa-Bianchi et al. (2006).
This measure takes
into account that after a hierarchical prediction rule makes its first error
in the top-down prediction process, it is not expected that the remaining predictions
at lower levels of the category tree are correct, which is why only the first
errors made should be penalized. Moreover, errors made in the upper levels of the
category tree should be penalized more strongly than errors occurring at the
lower levels. This loss measure particularly rewards early stoppings without
errors. See the 'Details' section for more details.
}
\note{
The score function calls \code{\link[=hloss]{hloss()}} from package \CRANpkg{hierclass}.

If the measure is undefined for the input, \code{NaN} is returned.
This can be customized by setting the field \code{na_value}.
}
\section{Dictionary}{

This \link{Measure} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary} \link{mlr_measures} or with the associated sugar function \code{\link[=msr]{msr()}}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{mlr_measures$get("classif.hloss")
msr("classif.hloss")
}\if{html}{\out{</div>}}
}

\section{Parameters}{
\tabular{llll}{
   Id \tab Type \tab Default \tab Range \cr
   w0 \tab numeric \tab - \tab \eqn{[0, 1]}{[0, 1]} \cr
}
}

\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[=Measure]{Measures}: \link{mlr_measures}

\code{as.data.table(mlr_measures)} for a complete table of all (also dynamically created) \link{Measure} implementations.

Other classification measures: 
\code{\link{mlr_measures_classif.hierfbeta}},
\code{\link{mlr_measures_classif.hierpr}},
\code{\link{mlr_measures_classif.hierre}},
\code{\link{mlr_measures_classif.spath}}

Other binary classification measures: 
\code{\link{mlr_measures_classif.hierfbeta}},
\code{\link{mlr_measures_classif.hierpr}},
\code{\link{mlr_measures_classif.hierre}},
\code{\link{mlr_measures_classif.spath}}
}
\concept{binary classification measures}
\concept{classification measures}
\keyword{datasets}
