% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr3_measures_hierclass.R
\docType{data}
\name{mlr_measures_classif.spath}
\alias{mlr_measures_classif.spath}
\alias{measure_spath}
\title{Shortest path loss measure}
\format{
An object of class \code{MeasureClassifHierarchical} (inherits from \code{MeasureClassif}, \code{Measure}, \code{R6}) of length 25.
}
\usage{
measure_spath
}
\description{
The shortest path loss measure. The unweighted shortest path loss measure,
introduced by Wang et al. (1999), is simply the distance between the true
and the predicted class in the category tree. The weighted shortest path
loss measure (Blockeel et al., 2002) is the measure by Wang et al. (1999) but with depth-specific
weights of the edges in the path, where smaller weights are used for lower
levels of the category tree. While the unweighted version is easier interpretable
it has the disadvantage that it does not take into account that
misclassifications at lower levels of the hierarchy are more forgivable than
those at upper levels.

In the weighted shortest path distance the weight assigned to an edge is
w0^l, where l is the depth of the edge (starting with 1 for the highest
edges in the category tree) and w0 is the input parameter \code{w0}.
}
\note{
The score function calls \code{\link[=spath]{spath()}} from package \CRANpkg{hierclass}.

If the measure is undefined for the input, \code{NaN} is returned.
This can be customized by setting the field \code{na_value}.
}
\section{Dictionary}{

This \link{Measure} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary} \link{mlr_measures} or with the associated sugar function \code{\link[=msr]{msr()}}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{mlr_measures$get("classif.spath")
msr("classif.spath")
}\if{html}{\out{</div>}}
}

\section{Parameters}{
\tabular{lllll}{
   Id \tab Type \tab Default \tab Levels \tab Range \cr
   type \tab character \tab - \tab weighted \tab - \cr
   w0 \tab numeric \tab - \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
}
}

\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[=Measure]{Measures}: \link{mlr_measures}

\code{as.data.table(mlr_measures)} for a complete table of all (also dynamically created) \link{Measure} implementations.

Other classification measures: 
\code{\link{mlr_measures_classif.hierfbeta}},
\code{\link{mlr_measures_classif.hierpr}},
\code{\link{mlr_measures_classif.hierre}},
\code{\link{mlr_measures_classif.hloss}}

Other binary classification measures: 
\code{\link{mlr_measures_classif.hierfbeta}},
\code{\link{mlr_measures_classif.hierpr}},
\code{\link{mlr_measures_classif.hierre}},
\code{\link{mlr_measures_classif.hloss}}
}
\concept{binary classification measures}
\concept{classification measures}
\keyword{datasets}
