% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr3_measures_hierclass.R
\name{mlr_measures_hierclass}
\alias{mlr_measures_hierclass}
\alias{MeasureClassifHierarchical}
\title{Hierarchical Performance/Loss Measures}
\description{
Collection of measures for hierarchical classification outcomes.
}
\examples{
# Load the required packages:
library("mlr3")
library("hierclass")

# Set seed for reproducibility:
set.seed(1234)

# Load example dataset:
data(datasim)

# Define the task for the top-down classification rule:
task = as_task_classif(ydepvar ~ ., data = datasim)

# Initialize the learner for the top-down classification rule:
learner = lrn("classif.topdown")

# Train a model of this learner for a subset of the task:
learner$train(task, row_ids = 1:300)
learner$model

# Obtain predictions for the remaining observations:
predictions = learner$predict(task, row_ids = 301:560)
predictions

# Measure the performance according to different performance/loss metrics:

# Hierarchical F-score:
measure = msr("classif.hierfbeta")
predictions$score(measure)

# Hierarchical precision:
measure = msr("classif.hierpr")
predictions$score(measure)

# Hierarchical recall:
measure = msr("classif.hierre")
predictions$score(measure)

# H-loss measure:
measure = msr("classif.hloss")
predictions$score(measure)

# Shortest path loss measure:
measure = msr("classif.spath")
predictions$score(measure)

# Classification accuracy (not a hierarchical performance/loss measure):
measure = msr("classif.acc")
predictions$score(measure)
}
\seealso{
\code{\link{hierre}}, \code{\link{hierpr}}, \code{\link{hierfbeta}}, \code{\link{hloss}}, \code{\link{spath}}
}
\author{
Florian Pfisterer
}
\section{Super classes}{
\code{\link[mlr3:Measure]{mlr3::Measure}} -> \code{\link[mlr3:MeasureClassif]{mlr3::MeasureClassif}} -> \code{MeasureClassifHierarchical}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fun}}{(`function`) \cr
Function that implements the corresponding metric.}

\item{\code{param_set}}{(`ParamSet`) \¢r
ParamSet defining available hyperparameters.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MeasureClassifHierarchical-new}{\code{MeasureClassifHierarchical$new()}}
\item \href{#method-MeasureClassifHierarchical-clone}{\code{MeasureClassifHierarchical$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="aggregate"><a href='../../mlr3/html/Measure.html#method-Measure-aggregate'><code>mlr3::Measure$aggregate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="format"><a href='../../mlr3/html/Measure.html#method-Measure-format'><code>mlr3::Measure$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="help"><a href='../../mlr3/html/Measure.html#method-Measure-help'><code>mlr3::Measure$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="print"><a href='../../mlr3/html/Measure.html#method-Measure-print'><code>mlr3::Measure$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="score"><a href='../../mlr3/html/Measure.html#method-Measure-score'><code>mlr3::Measure$score()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasureClassifHierarchical-new"></a>}}
\if{latex}{\out{\hypertarget{method-MeasureClassifHierarchical-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureClassifHierarchical$new(fun, param_set = paradox::ps())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fun}}{(`function`) \cr
Function that implements the corresponding metric.}

\item{\code{param_set}}{(`ParamSet`) \¢r
ParamSet defining available hyperparameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasureClassifHierarchical-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MeasureClassifHierarchical-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureClassifHierarchical$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
